example_seven/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
asyncpg==0.31.0
attrs==25.4.0
banks==2.2.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
ddgs>=0.1.0
email-validator==2.3.0
eval_type_backport==0.2.2
exceptiongroup==1.3.1
fastapi==0.123.2
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.2
fastar==0.8.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.2.4
griffe==1.14.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.8
llama-index-core==0.14.8
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-openai==0.6.10
llama-index-readers-file==0.5.5
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-postgres==0.7.2
llama-index-workflows==2.11.5
llama-parse==0.6.54
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.2
numpy==2.0.2
openai==2.8.1
packaging==25.0
pandas==2.2.3
pgvector==0.4.1
pillow==11.3.0
platformdirs==4.4.0
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pypdf==6.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
sentry-sdk==2.46.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.49.3
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.12.0
tomli==2.3.0
tqdm==4.67.1
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0

